+++ 02/lib.h
@@ -1,2 +1,4 @@
-long part1(const char* path);
-long part2(const char* path);
+#include <cstdint>
+
+int32_t part1(const char* path);
+int32_t part2(const char* path);
